backend/requirements.txt
pdfplumber
nltk
flask
transformers
torch
sentence-transformers
faiss-cpu




